+++ CHANGELOG.md
@@ -62,6 +62,9 @@ This is a major release, and it might be not compatible with your current usage 
 -   `<Modal />`:
     -   Add ModalContext to track open/close state of all used application modals.
     -   Add `modalId` property to give a modal a unique ID for tracking purposes.
+    -   `preventReactFlowEvents`: Adds 'nopan', 'nowheel' and 'nodrag' classes to Modal's overlay classes in order to prevent react-flow to react to drag and pan actions in modals.
+        This is enabled by default for <SimpleDialog />.
+
 
 ### Removed
 
@@ -89,8 +92,6 @@ This is a major release, and it might be not compatible with your current usage 
 
 ### Fixed
 
--   `<Modal />`:
-    -   Add 'nopan', 'nowheel' and 'nodrag' classes to Modal's overlay classes in order to always prevent react-flow to react to drag and pan actions in modals.
 -   `<CodeAutocompleteField />`:
     -   In multiline mode, validation errors might be highlighted incorrectly (relative line offset added).